[MOVE/#390] 결제 패키지 다이얼로그 묶음
