+++ app/src/main/java/com/el/yello/presentation/main/MainActivity.kt
@@ -22,7 +22,7 @@ import com.el.yello.presentation.main.myyello.read.MyYelloReadActivity
 import com.el.yello.presentation.main.profile.info.ProfileFragment
 import com.el.yello.presentation.main.recommend.RecommendFragment
 import com.el.yello.presentation.main.yello.YelloFragment
-import com.el.yello.presentation.pay.PayReSubsNoticeDialog
+import com.el.yello.presentation.pay.dialog.PayReSubsNoticeDialog
 import com.el.yello.presentation.util.dp
 import com.el.yello.util.amplitude.AmplitudeUtils
 import com.el.yello.util.context.yelloSnackbar
